+++ timezones/utc/utc_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/matthalp/go-meridian/v2/timezones/pt"
+	"github.com/matthalp/go-meridian/v3/timezones/pt"
 )
 
 func TestUTCLocation(t *testing.T) {